Raise SPU voices to maximum gain
The preceding dry-safe limit left the output below the requested level. Allow overlapping voices to clip while keeping the full envelope and mix ranges valid at the SPU volume ceiling.

+++ tests/audio_fixture.c
@@ -135,7 +135,8 @@ static void synthesis_checks(void) {
                 && SPU_CH_LOOP_ADDR(i)==SPU_CH_LOOP_ADDR(i+1);
         // Capture is pre-volume voice output, not the final mixer. The bound
         // scales its observed peak by all twelve complementary gain budgets.
-        log_message("WAVE wave=%d peak=%d pairs=%d bound=%d\n",wave,peak,pairs,(peak*SEQUENCER_VOICES*AUDIO_LEVEL+16383)/16384);
+        log_message("WAVE wave=%d peak=%d pairs=%d bound=%d\n",wave,peak,pairs,
+            (int)(((uint64_t)peak*SEQUENCER_VOICES*AUDIO_LEVEL+16383)/16384));
         report("wave chord");
         audio_platform_update(&editor.score,0,0); frames(2);
     }
